requirements.txt
Flask==3.0.3
scikit-learn==1.6.0
pandas==2.2.2
numpy==2.1.0
matplotlib==3.9.2
google-generativeai==0.8.3
